Moved all commands to bdaybot commands to "bdaybot_commands.py"
